+++ src/app/components/cadastro-gerencia/semestre/gerencia-semestre/gerencia-semestre.component.ts
@@ -9,6 +9,7 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatDialog } from '@angular/material/dialog';
 import { Semestre } from '../../../../models/Semestre';
 import { EdicaoSemestreComponent } from '../edicao-semestre/edicao-semestre.component';
+import { ModalDialogOkComponent } from '../../../modal-dialog/modal-dialog-ok/modal-dialog-ok.component';
 
 @Component({
   selector: 'app-gerencia-semestre',
@@ -62,16 +63,66 @@ export class GerenciaSemestreComponent implements OnInit {
     });
   }
 
+  // excluir(semestre: Semestre): void {
+  //   const confirmacao = confirm('Tem certeza que deseja excluir este semestre?');
+  //   if (confirmacao) {
+  //     this.service.remove(semestre._id).subscribe(() => {
+  //       this.semestres = this.semestres.filter(e => e._id !== semestre._id);
+  //       this.onSucess(true);
+  //     }, error => {
+  //       this.onFailed();
+  //     });
+  //   }
+  // }
+
   excluir(semestre: Semestre): void {
-    const confirmacao = confirm('Tem certeza que deseja excluir este semestre?');
-    if (confirmacao) {
-      this.service.remove(semestre._id).subscribe(() => {
-        this.semestres = this.semestres.filter(e => e._id !== semestre._id);
-        this.onSucess(true);
-      }, error => {
-        this.onFailed();
-      });
-    }
+    this.service.getRegistrosUsandoSemestre(semestre._id).subscribe(registros => {
+      if (registros.length > 0) {
+        this.mostrarMensagemErro(registros); 
+      } else {
+        const confirmacao = confirm('Tem certeza que deseja excluir este semestre?');
+        if (confirmacao) {
+          this.service.remove(semestre._id).subscribe(() => {
+            this.semestres = this.semestres.filter(e => e._id !== semestre._id);
+            this.onSucess(true);
+          }, error => {
+            this.onFailed();
+          });
+        }
+      }
+    });
+  }
+
+  mostrarMensagemErro(registros: any[]): void {
+    let cont = 1;
+
+    const itensLista = registros.map(registro => {
+        return `<span>Aula ${cont++}</span> <br>
+                Disciplina: ${registro.discipline.name} - (${registro.discipline.course.name}) <br>
+                Professor(a): ${registro.teacher.name} <br>
+        
+        `;
+    });
+
+    const dialogDataForm = {
+      title: 'Erro ao Excluir Semestre',
+      message: `
+    <div mat-dialog-content>
+      <p>Exclua os seguintes registros primeiramente:</p>
+      <span>Total de Aulas: ${itensLista.length}</span>
+      <ul> 
+
+        ${itensLista.map(item => `<li>${item}</li><br>`).join('')}
+        
+      </ul>
+    </div>
+  `,
+    };
+
+    this.dialog.open(ModalDialogOkComponent, {
+      data: dialogDataForm,
+      backdropClass: 'backdrop'
+    });
   }
 
   onFailed() {
